🎨 Create ./tests/helpers/ directory that defines variables and functions used in testing
