+++ package.json
@@ -30,6 +30,7 @@
     "lint:eslint": "eslint --ext .ts,.js ./",
     "lint:tsc-src": "tsc --noEmit",
     "lint:tsc-test": "tsc -p ./tests/ --noEmit",
+    "lint:tsc-test-helpers": "tsc -p ./tests/helpers/ --noEmit",
     "release": "run-s release:validation release:build release:publish",
     "release:build": "run-s clean build",
     "release:publish": "run-s release:publish:git",
